+++ Point3.hh
@@ -7,15 +7,15 @@ class Point3
 {
 public:
     Point3();
-    Point3(float x, float y, float z);
+    Point3(double x, double y, double z);
     Point3 operator+=(const Point3& p);
   Point3 operator+(const Point3& p) const;
   Point3 operator-(const Point3& p) const;
 
 
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
 };
 
 std::ostream& operator<<(std::ostream &out, const Point3& point);
